move animation into own module
